Added test suite for bringing up agora and embedded mongo before all tests and tearing it down after all tests.
Removing mongodb service from travis yaml.
