+++ .travis.yml
@@ -1,8 +1,6 @@
 language: scala
 scala:
   - 2.11.6
-services:
-  - mongodb
 jdk:
   - oraclejdk8
 script: sbt clean coverage test
